feat: Updates ignore patterns in biome.json and vite.config.js, fixes types in Text component, and updates package.json
