+++ src/consts/colors.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import React from "react";
+import { twMerge } from "tailwind-merge";
 import { COLOR_SCHEME } from ".";
 
 const meta = {
@@ -16,47 +16,55 @@ type Story = StoryObj<typeof meta>;
 export const Colors: Story = {
 	render: () => (
 		<div>
-			{Object.keys(COLOR_SCHEME).map((color) => (
-				<div key={color} className="grid gap-8 grid-cols-2">
-					<span className="text-4xl font-normal">
-						{color[0].toUpperCase() + color.slice(1)}
-					</span>
-					<div className="grid gap-2">
-						{Object.keys(COLOR_SCHEME[color]).map((shade) =>
-							shade === "purple" && color === "primary" ? (
-								<>
+			{(Object.keys(COLOR_SCHEME) as (keyof typeof COLOR_SCHEME)[]).map(
+				(color) => (
+					<div key={color} className="grid gap-8 grid-cols-2">
+						<span className="text-4xl font-normal">
+							{color[0].toUpperCase() + color.slice(1)}
+						</span>
+						<div className="grid gap-2">
+							{Object.keys(
+								COLOR_SCHEME[
+									color
+								] as (typeof COLOR_SCHEME)[keyof typeof COLOR_SCHEME],
+							).map((shade) =>
+								shade === "purple" && color === "primary" ? (
+									<>
+										<div
+											key={`${color + shade}50`}
+											className={"rounded-2xl text-center p-5"}
+											style={{
+												backgroundColor: COLOR_SCHEME[color][shade][50],
+											}}
+										>
+											<span className="text-sm text-black-100">{shade} 50</span>
+										</div>
+										<div
+											key={`${color + shade}DEFAULT`}
+											className={"rounded-2xl text-center p-5"}
+											style={{
+												backgroundColor: COLOR_SCHEME[color][shade].DEFAULT,
+											}}
+										>
+											<span className="text-sm text-black-100">{shade}</span>
+										</div>
+									</>
+								) : (
 									<div
-										key={`${color + shade}50`}
-										className={"rounded-2xl text-center p-5"}
-										style={{
-											backgroundColor: COLOR_SCHEME[color][shade][50],
-										}}
-									>
-										<span className="text-sm text-black-100">{shade} 50</span>
-									</div>
-									<div
-										key={`${color + shade}DEFAULT`}
-										className={"rounded-2xl text-center p-5"}
-										style={{
-											backgroundColor: COLOR_SCHEME[color][shade].DEFAULT,
-										}}
+										key={color + shade}
+										className={twMerge(
+											"rounded-2xl text-center p-5",
+											`bg-${color}-${shade}`,
+										)}
 									>
 										<span className="text-sm text-black-100">{shade}</span>
 									</div>
-								</>
-							) : (
-								<div
-									key={color + shade}
-									className={"rounded-2xl text-center p-5"}
-									style={{ backgroundColor: COLOR_SCHEME[color][shade] }}
-								>
-									<span className="text-sm text-black-100">{shade}</span>
-								</div>
-							),
-						)}
+								),
+							)}
+						</div>
 					</div>
-				</div>
-			))}
+				),
+			)}
 		</div>
 	),
 };
